same for hex upper

+++ parser.c
@@ -13,7 +13,7 @@ int	call_printer(t_tokens *tokens, va_list args_list)
 	else if (tokens->type ==  STR)
 		ret_val = print_string_bonus(va_arg(args_list, char *), tokens);
 	else if (tokens->type ==  DECIM || tokens->type == INT)
-		ret_val = print_integers_bonus(va_arg(args_list, int), tokens);
+		ret_val = print_integers_bonus(va_arg(args_list, long), tokens);
 	else if (tokens->type == PRCND)
 		ret_val = print_char('%');
 	else if (tokens->type == UNSI)
